requirements.txt
playwright
pyyaml
loguru
ttkbootstrap
fastapi
uvicorn
openai>=1.0.0
